@@ -1,5 +1,5 @@
 import { registerLocaleData } from '@angular/common';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import localeDE from '@angular/common/locales/de';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { MatBottomSheetModule } from '@angular/material/bottom-sheet';
@@ -57,68 +57,62 @@ import { ConfirmChoiceComponent } from './vote/vote-form/confirm-choice/confirm-
 
 registerLocaleData(localeDE);
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    ConditionsComponent,
-    HomepageComponent,
-    ImprintComponent,
-    PrivacyComponent,
-    MainNavComponent,
-    ContactComponent,
-    LoginComponent,
-    RegisterComponent,
-    ForgotPasswordComponent,
-    ResetPasswordComponent,
-    RegisterComponent,
-    MainNavComponent,
-    EditProfileComponent,
-    ChangePasswordComponent,
-    DeleteProfileComponent,
-    ViewProfileComponent,
-    CookieBannerComponent,
-    ConfirmAccountInstructionComponent,
-    ConfirmAccountComponent,
-    ResendAccountConfirmationEmailComponent,
-    ConfirmChoiceComponent,
-    FeedbackComponent,
-    VersionComponent,
-    FeedbackDialogComponent,
-    EmailNotificationSettingsComponent,
-    ContactListComponent,
-    AddContactComponent,
-    GroupDisplayComponent,
-    EditGroupComponent,
-    ChangeNoteComponent,
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    MatToolbarModule,
-    MatSidenavModule,
-    MatNativeDateModule,
-    HttpClientModule,
-    MatSnackBarModule,
-    MatBottomSheetModule,
-    MatDialogModule,
-    StoreModule.forRoot({}),
-    StoreModule.forFeature(authFeature),
-    StoreModule.forFeature(contactsFeature),
-    StoreModule.forFeature(cookieFeature),
-    StoreModule.forFeature(emailNotificationSettingsFeature),
-    StoreModule.forFeature(notificationFeature),
-    EffectsModule.forRoot(appEffects),
-    ...environment.devModules,
-    AppCommonModule,
-    MtxNativeDatetimeModule,
-  ],
-  providers: [
-    {provide: MAT_DATE_LOCALE, useValue: 'de-DE'},
-    {provide: DateAdapter, useClass: EUDateAdapter},
-    {provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true},
-  ],
-  bootstrap: [AppComponent],
-  schemas: [CUSTOM_ELEMENTS_SCHEMA],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        ConditionsComponent,
+        HomepageComponent,
+        ImprintComponent,
+        PrivacyComponent,
+        MainNavComponent,
+        ContactComponent,
+        LoginComponent,
+        RegisterComponent,
+        ForgotPasswordComponent,
+        ResetPasswordComponent,
+        RegisterComponent,
+        MainNavComponent,
+        EditProfileComponent,
+        ChangePasswordComponent,
+        DeleteProfileComponent,
+        ViewProfileComponent,
+        CookieBannerComponent,
+        ConfirmAccountInstructionComponent,
+        ConfirmAccountComponent,
+        ResendAccountConfirmationEmailComponent,
+        ConfirmChoiceComponent,
+        FeedbackComponent,
+        VersionComponent,
+        FeedbackDialogComponent,
+        EmailNotificationSettingsComponent,
+        ContactListComponent,
+        AddContactComponent,
+        GroupDisplayComponent,
+        EditGroupComponent,
+        ChangeNoteComponent,
+    ],
+    bootstrap: [AppComponent],
+    schemas: [CUSTOM_ELEMENTS_SCHEMA], imports: [BrowserModule,
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        MatToolbarModule,
+        MatSidenavModule,
+        MatNativeDateModule,
+        MatSnackBarModule,
+        MatBottomSheetModule,
+        MatDialogModule,
+        StoreModule.forRoot({}),
+        StoreModule.forFeature(authFeature),
+        StoreModule.forFeature(contactsFeature),
+        StoreModule.forFeature(cookieFeature),
+        StoreModule.forFeature(emailNotificationSettingsFeature),
+        StoreModule.forFeature(notificationFeature),
+        EffectsModule.forRoot(appEffects),
+        ...environment.devModules,
+        AppCommonModule,
+        MtxNativeDatetimeModule], providers: [
+        { provide: MAT_DATE_LOCALE, useValue: 'de-DE' },
+        { provide: DateAdapter, useClass: EUDateAdapter },
+        { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule {}
