@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 
 import { SurveyService } from '../survey.service';
 
@@ -11,6 +11,7 @@ import { SurveyPreviewComponent } from './survey-preview/survey-preview.componen
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('SurveyCreateComponent', () => {
   let component: SurveyCreateComponent;
@@ -18,10 +19,10 @@ describe('SurveyCreateComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ SurveyCreateComponent, SurveyCreateFormComponent, SurveyPreviewComponent ],
-      imports: [ HttpClientTestingModule, RouterTestingModule, MatStepperModule, NoopAnimationsModule, MatButtonModule, MatCardModule ],
-      providers: [ SurveyService ],
-    })
+    declarations: [SurveyCreateComponent, SurveyCreateFormComponent, SurveyPreviewComponent],
+    imports: [RouterTestingModule, MatStepperModule, NoopAnimationsModule, MatButtonModule, MatCardModule],
+    providers: [SurveyService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+})
     .compileComponents();
   });
 
