@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatButtonModule } from '@angular/material/button';
 import { MatExpansionModule } from '@angular/material/expansion';
@@ -9,6 +9,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { SurveyService } from '../survey.service';
 
 import { SurveyResultsComponent } from './survey-results.component';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('SurveyResultsComponent', () => {
   let component: SurveyResultsComponent;
@@ -16,10 +17,10 @@ describe('SurveyResultsComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ SurveyResultsComponent ],
-      imports: [ RouterTestingModule, HttpClientTestingModule, MatTableModule, MatExpansionModule, MatProgressSpinnerModule, MatButtonModule, MatIconModule ],
-      providers: [ SurveyService ],
-    })
+    declarations: [SurveyResultsComponent],
+    imports: [RouterTestingModule, MatTableModule, MatExpansionModule, MatProgressSpinnerModule, MatButtonModule, MatIconModule],
+    providers: [SurveyService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+})
     .compileComponents();
   });
 
