@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
@@ -8,6 +8,7 @@ import { RouterTestingModule } from '@angular/router/testing';
 
 import { SurveyService } from '../survey/survey.service';
 import { VoteListComponent } from './vote-list.component';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('SurveyListComponent', () => {
   let component: VoteListComponent;
@@ -15,10 +16,10 @@ describe('SurveyListComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ VoteListComponent ],
-      imports: [ RouterTestingModule, HttpClientTestingModule, MatCardModule, MatButtonModule, MatProgressSpinnerModule, MatIconModule ],
-      providers: [ SurveyService ]
-    })
+    declarations: [VoteListComponent],
+    imports: [RouterTestingModule, MatCardModule, MatButtonModule, MatProgressSpinnerModule, MatIconModule],
+    providers: [SurveyService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+})
       .compileComponents();
   });
 
